Fix PackIt's tarball name to match tito's tarball name
Tito creates a tarball/rpm that says 'restraint-0.4.11', whereas PackIt
uses <name>-<git tag> and creates 'restraint-restraint.0.4.11'.  This
makes it hard to 'dnf update restraint' to use PackIt's rpm on top of
the official Tito built rpm because restraint-0.4.11 >
restraint-restraint.0.4.11.<packit extra>.prxxx.

Let's align the tarballs so we can use PackIt's rpm as a natural dnf
update to the one in the provisioner allowing for CI tests.

The simplest solution is to override PackIt's get-current-version action
with our custom one.

Signed-off-by: Don Zickus <dzickus@redhat.com>